// src/models/Building.ts
export interface Building {
  regionId: number;
  farmland: number;
  lightFactory: number;
  buildingDepartment: number;
  ironMine: number;
  oilWell: number;
  coalMine: number;
  rareEarthMine: number;
  rareMetalMine: number;
  steelPlant: number;
  powerPlant: number;
  hydroelectricPowerStation: number;
  militaryFactory: number;
  mechanicalFactory: number;
}